requirements.txt
qrcode
Pillow
pyzbar
pytest
hashlib
hmac